fix issue with image pull secrets for workflows
