+++ pkg/imageinspector/skopeofetcher.go
@@ -19,6 +19,10 @@ func NewSkopeoFetcher() InfoFetcher {
 }
 
 func (s *skopeoFetcher) Fetch(ctx context.Context, registry, image string, pullSecrets []corev1.Secret) (*Info, error) {
+	// If registry is not provided, extract it from the image name
+	if registry == "" {
+		registry = extractRegistry(image)
+	}
 	client, err := skopeo.NewClientFromSecrets(pullSecrets, registry)
 	if err != nil {
 		return nil, err
@@ -37,6 +41,22 @@ func (s *skopeoFetcher) Fetch(ctx context.Context, registry, image string, pullS
 		User:       user,
 		Group:      group,
 	}, nil
+}
+
+// extractRegistry takes a container image string and returns the registry part.
+// It defaults to "docker.io" if no registry is specified.
+func extractRegistry(image string) string {
+	defaultRegistry := "https://index.docker.io/v1/"
+	parts := strings.Split(image, "/")
+	// If the image is just a name, return the default registry.
+	if len(parts) == 1 {
+		return defaultRegistry
+	}
+	// If the first part contains '.' or ':', it's likely a registry.
+	if strings.Contains(parts[0], ".") || strings.Contains(parts[0], ":") {
+		return parts[0]
+	}
+	return defaultRegistry
 }
 
 func determineUserGroupPair(userGroupStr string) (int64, int64) {
