+++ pkg/skopeo/client.go
@@ -86,14 +86,16 @@ func (c *client) Inspect(registry, image string) (*DockerImage, error) {
 	}
 
 	if len(c.dockerAuthConfigs) != 0 {
+		// TODO: Is it a good idea to randomly select a secret?
 		i := rand.Intn(len(c.dockerAuthConfigs))
 		args = append(args, "--creds", c.dockerAuthConfigs[i].Username+":"+c.dockerAuthConfigs[i].Password)
 	}
 
-	config := "docker://" + image
-	if registry != "" {
-		config = registry + "/" + image
+	// If registry is provided via config and the image does not start with the registry, prepend it
+	if registry != "" && !strings.HasPrefix(image, registry) {
+		image = registry + "/" + image
 	}
+	config := "docker://" + image
 
 	args = append(args, "--config", config)
 	result, err := process.Execute("skopeo", args...)
@@ -157,10 +159,6 @@ func ParseSecretData(imageSecrets []corev1.Secret, registry string) ([]DockerAut
 			return nil, fmt.Errorf("imagePullSecret %s contains neither .dockercfg nor .dockerconfigjson", imageSecret.Name)
 		}
 
-		// If registry is not provided, extract it from the image name
-		if registry == "" {
-			registry = extractRegistry(imageSecret.Name)
-		}
 		// Determine if there is a secret for the specified registry
 		if creds, ok := auths.Auths[registry]; ok {
 			username, password, err := extractRegistryCredentials(creds)
@@ -169,28 +167,10 @@ func ParseSecretData(imageSecrets []corev1.Secret, registry string) ([]DockerAut
 			}
 
 			results = append(results, DockerAuthConfig{Username: username, Password: password})
-		} else {
-			return nil, fmt.Errorf("secret %s is not defined for registry: %s", imageSecret.Name, registry)
 		}
 	}
 
 	return results, nil
-}
-
-// extractRegistry takes a container image string and returns the registry part.
-// It defaults to "docker.io" if no registry is specified.
-func extractRegistry(image string) string {
-	defaultRegistry := "https://index.docker.io/v1/"
-	parts := strings.Split(image, "/")
-	// If the image is just a name, return the default registry.
-	if len(parts) == 1 {
-		return defaultRegistry
-	}
-	// If the first part contains '.' or ':', it's likely a registry.
-	if strings.Contains(parts[0], ".") || strings.Contains(parts[0], ":") {
-		return parts[0]
-	}
-	return defaultRegistry
 }
 
 func extractRegistryCredentials(creds DockerAuthConfig) (username, password string, err error) {
